Get Latest

+++ OfficeDevPnP.Core/OfficeDevPnP.Core.Tests/Framework/ObjectHandlers/ObjectTermGroupsTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Linq;
 using Microsoft.SharePoint.Client;
 using Microsoft.SharePoint.Client.Taxonomy;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -21,30 +22,40 @@ namespace OfficeDevPnP.Core.Tests.Framework.ObjectHandlers
         [TestInitialize]
         public void Initialize()
         {
-            _termSetGuid = Guid.NewGuid();
-            _termGroupGuid = Guid.NewGuid();
+            if (!TestCommon.AppOnlyTesting())
+            {
+                _termSetGuid = Guid.NewGuid();
+                _termGroupGuid = Guid.NewGuid();
+            }
+            else
+            {
+                Assert.Inconclusive("Taxonomy tests are not supported when testing using app-only");
+            }
         }
 
         [TestCleanup]
         public void CleanUp()
         {
-            using (var ctx = TestCommon.CreateClientContext())
+            if (!TestCommon.AppOnlyTesting())
             {
-                try
+                using (var ctx = TestCommon.CreateClientContext())
                 {
-                    TaxonomySession session = TaxonomySession.GetTaxonomySession(ctx);
+                    try
+                    {
+                        TaxonomySession session = TaxonomySession.GetTaxonomySession(ctx);
 
-                    var store = session.GetDefaultSiteCollectionTermStore();
-                    var termSet = store.GetTermSet(_termSetGuid);
-                    termSet.DeleteObject();
+                        var store = session.GetDefaultSiteCollectionTermStore();
+                        var termSet = store.GetTermSet(_termSetGuid);
+                        termSet.DeleteObject();
 
-                    var termGroup = store.GetGroup(_termGroupGuid);
-                    termGroup.DeleteObject();
-                    store.CommitAll();
-                    ctx.ExecuteQueryRetry();
-                }
-                catch
-                {
+                        var termGroup = store.GetGroup(_termGroupGuid);
+                        termGroup.DeleteObject();
+                        store.CommitAll();
+                        ctx.ExecuteQueryRetry();
+                    }
+                    catch
+                    {
+                    }
                 }
             }
         }
@@ -58,7 +69,7 @@ namespace OfficeDevPnP.Core.Tests.Framework.ObjectHandlers
 
             List<TermSet> termSets = new List<TermSet>();
 
-            TermSet termSet = new TermSet(_termSetGuid, "TestProvisioningTermSet", null, null);
+            TermSet termSet = new TermSet(_termSetGuid, "TestProvisioningTermSet", null, true, false, null, null);
 
             List<Term> terms = new List<Term>();
 
@@ -85,7 +96,7 @@ namespace OfficeDevPnP.Core.Tests.Framework.ObjectHandlers
             {
                 TokenParser.Initialize(ctx.Web, template);
 
-                new ObjectTermGroups().ProvisionObjects(ctx.Web, template);
+                new ObjectTermGroups().ProvisionObjects(ctx.Web, template, new ProvisioningTemplateApplyingInformation());
 
                 TaxonomySession session = TaxonomySession.GetTaxonomySession(ctx);
 
@@ -100,10 +111,21 @@ namespace OfficeDevPnP.Core.Tests.Framework.ObjectHandlers
                 ctx.ExecuteQueryRetry();
 
                 Assert.IsInstanceOfType(group, typeof(Microsoft.SharePoint.Client.Taxonomy.TermGroup));
-                Assert.IsInstanceOfType(set, typeof (Microsoft.SharePoint.Client.Taxonomy.TermSet));
+                Assert.IsInstanceOfType(set, typeof(Microsoft.SharePoint.Client.Taxonomy.TermSet));
                 Assert.IsTrue(set.Terms.Count == 2);
 
+
+                var creationInfo = new ProvisioningTemplateCreationInformation(ctx.Web) { BaseTemplate = ctx.Web.GetBaseTemplate() };
+
+                var template2 = new ProvisioningTemplate();
+                template2 = new ObjectTermGroups().ExtractObjects(ctx.Web, template, creationInfo);
+
+                Assert.IsTrue(template.TermGroups.Any());
+                Assert.IsInstanceOfType(template.TermGroups, typeof(List<TermGroup>));
             }
+
+
         }
+
     }
 }
